refactor(workspace): type decoded FITS geometry as the lengths it is (#870)
* refactor(workspace): type decoded FITS geometry as the lengths it is

`rp_fits::reader` parsed NAXIS from `i64` into a `u32` that every
caller immediately widened back to `usize` to index the buffer it
describes. Returning `usize` deletes the middle step rather than
respelling it: one fallible conversion instead of one fallible plus
one lossy, and the ~26 `as usize` casts downstream stop existing.

The rule the sites were really asking for is where the conversion
belongs, not how to spell it: `usize` must never appear in a
serialized format, so anything bound for disk or a wire keeps a fixed
width and anything that indexes a buffer is a `usize`. That splits
`rp-fits` cleanly — the reader hands back a shape, the writer takes
header cards — and settles the callers the same way. Alpaca
`NumX`/`StartX`, the `ImageBytes` header and a PNG's dimensions stay
`u32`; `crop_subframe`, `Array2::from_shape_vec` and the preview
subsampling become `usize`.

The three boundary conversions that survive fold into an error the
function already returns: a subframe too large for a `usize` cannot
fit the source buffer either, so it lands in the existing bounds
check instead of earning a variant of its own.

Production `x as usize` sites: 76 -> 50.

* fix(rp-fits,sky-survey-camera): reject a frame whose data is short of its NAXIS

`read_primary` never checked the decoded pixel count against
`NAXIS1 × NAXIS2`, so a truncated stream produced a `FitsImage` whose
geometry over-promised its buffer. Most consumers survived that by
accident — `Array2::from_shape_vec` and the preview's own guard both
reject a mismatch — but `crop_subframe` walks the buffer a row at a
time and would slice past the end:

    let start = row * src_w + sx;
    out.extend_from_slice(&src[start..start + nx]);

Its bounds check compares the subframe against the *declared* size, not
against `src.len()`, so the panic was one short survey response away —
and sky-survey-camera decodes FITS fetched over HTTP.

Checked at the root, where the invariant belongs: `FitsImage`'s geometry
now describes its buffer because `read_primary` refuses to build one
where it does not. `crop_subframe` keeps a guard of its own rather than
slicing on a guarantee made in another crate.

Also reorders the `ImageBytes` width check in `disk_resolve` ahead of
the narrowing pass, so an oversize frame is dropped before its pixels
are walked.

Both found in review of #870.

+++ services/sky-survey-camera/src/fits.rs
@@ -15,10 +15,12 @@ use rp_fits::reader::read_primary_as_i32;
 
 pub use rp_fits::FitsError;
 
+/// `width` and `height` are `usize`: they describe `data`, and every
+/// consumer uses them to slice or shape it.
 #[derive(Debug, Clone)]
 pub struct FitsImage {
-    pub width: u32,
-    pub height: u32,
+    pub width: usize,
+    pub height: usize,
     pub data: Vec<i32>,
 }
